+++ src/test/java/model/Hour.java
@@ -6,33 +6,33 @@ import java.util.List;
 
 public class Hour {
    private long dt;
-   private int temp;
+   private double temp;
 
    @JsonProperty ("feels_like")
-   private int feelsLike;
+   private double feelsLike;
 
    private int pressure;
    private int humidity;
 
    @JsonProperty ("dew_point")
-   private int dewPoint;
+   private double dewPoint;
 
-   private int uvi;
+   private double uvi;
    private int clouds;
    private int visibility;
 
     @JsonProperty("wind_speed")
-    private int windSpeed;
+    private double windSpeed;
 
     @JsonProperty("wind_deg")
     private int windDeg;
 
     @JsonProperty("wind_gust")
-    private int windGust;
+    private double windGust;
 
     private List<Weather> weather;
 
-    private int pop;
+    private double pop;
 
 
     public long getDt() {
@@ -43,19 +43,19 @@ public class Hour {
         this.dt = dt;
     }
 
-    public int getTemp() {
+    public double getTemp() {
         return temp;
     }
 
-    public void setTemp(int temp) {
+    public void setTemp(double temp) {
         this.temp = temp;
     }
 
-    public int getFeelsLike() {
+    public double getFeelsLike() {
         return feelsLike;
     }
 
-    public void setFeelsLike(int feelsLike) {
+    public void setFeelsLike(double feelsLike) {
         this.feelsLike = feelsLike;
     }
 
@@ -75,19 +75,19 @@ public class Hour {
         this.humidity = humidity;
     }
 
-    public int getDewPoint() {
+    public double getDewPoint() {
         return dewPoint;
     }
 
-    public void setDewPoint(int dewPoint) {
+    public void setDewPoint(double dewPoint) {
         this.dewPoint = dewPoint;
     }
 
-    public int getUvi() {
+    public double getUvi() {
         return uvi;
     }
 
-    public void setUvi(int uvi) {
+    public void setUvi(double uvi) {
         this.uvi = uvi;
     }
 
@@ -107,11 +107,11 @@ public class Hour {
         this.visibility = visibility;
     }
 
-    public int getWindSpeed() {
+    public double getWindSpeed() {
         return windSpeed;
     }
 
-    public void setWindSpeed(int windSpeed) {
+    public void setWindSpeed(double windSpeed) {
         this.windSpeed = windSpeed;
     }
 
@@ -123,11 +123,11 @@ public class Hour {
         this.windDeg = windDeg;
     }
 
-    public int getWindGust() {
+    public double getWindGust() {
         return windGust;
     }
 
-    public void setWindGust(int windGust) {
+    public void setWindGust(double windGust) {
         this.windGust = windGust;
     }
 
@@ -139,11 +139,11 @@ public class Hour {
         this.weather = weather;
     }
 
-    public int getPop() {
+    public double getPop() {
         return pop;
     }
 
-    public void setPop(int pop) {
+    public void setPop(double pop) {
         this.pop = pop;
     }
 }
